fix: fix user profile pic change
