+++ src/Components/Comment/Comment.jsx
@@ -1,15 +1,27 @@
-import React from 'react';
+import React, { useEffect, useState } from 'react';
 import { Avatar, Box, Heading, HStack, IconButton, Text } from '@chakra-ui/react';
 import { FaEllipsisV } from 'react-icons/fa';
 import { useSelector } from 'react-redux';
 
-const Comment = ({ firstName, lastName, profilePic, text, username, _id }) => {
+const Comment = ({ firstName, lastName, profilePic, text, username, _id, img }) => {
+	const [commentImg, setCommentImg] = useState();
 	const { user } = useSelector((state) => state.auth);
+
+	useEffect(() => {
+		if (username === user.username) {
+			setCommentImg(img);
+		}
+	}, [img, user.username, username]);
 
 	return (
 		<>
 			<HStack mt="3" gap="2" align="flex-start">
-				<Avatar name={`${firstName} ${lastName}`} src={profilePic} size="sm" fontSize="10px" />
+				<Avatar
+					name={`${firstName} ${lastName}`}
+					src={commentImg || profilePic}
+					size="sm"
+					fontSize="10px"
+				/>
 				<Box bgColor="gray.200" p="3" flexGrow="1" borderRadius="lg" position="relative">
 					<Heading as="h4" fontWeight="700" fontSize="16" mb="2">
 						{`${firstName} ${lastName}`}{' '}
