+++ src/Redux/Slice/usersSlice.js
@@ -1,5 +1,5 @@
 import { createSlice } from '@reduxjs/toolkit';
-import { editProfile, getUsers } from 'Redux/Thunk';
+import { getUsers } from 'Redux/Thunk';
 
 const initialState = {
 	users: [],
@@ -20,19 +20,6 @@ const usersSlice = createSlice({
 		},
 		[getUsers.rejected]: (state, action) => {
 			state.isLoading = false;
-		},
-		[editProfile.pending]: (state, action) => {
-			state.isLoading = true;
-		},
-		[editProfile.fulfilled]: (state, { payload }) => {
-			state.isLoading = false;
-			state.users = state.users.map((user) =>
-				user.username === payload.data.user.username ? payload.data.user : user
-			);
-		},
-		[editProfile.rejected]: (state, action) => {
-			state.isLoading = false;
-			console.error(action);
 		},
 	},
 });
