+++ src/Redux/Thunk/authThunk.js
@@ -20,3 +20,21 @@ export const signup = createAsyncThunk('auth/signup', async (user, { rejectWithV
 		return rejectWithValue({ status: error.response.status, data: error.response.data });
 	}
 });
+
+export const editProfile = createAsyncThunk(
+	'auth/editProfile',
+	async ({ userData, token }, { rejectWithValue }) => {
+		try {
+			const response = await axios.post(
+				'/api/users/edit',
+				{
+					userData,
+				},
+				{ headers: { authorization: token } }
+			);
+			return { data: response.data, status: response.status };
+		} catch (error) {
+			return rejectWithValue({ status: error.response.status, data: error.response.data });
+		}
+	}
+);
