+++ src/Redux/Thunk/usersThunk.js
@@ -6,21 +6,3 @@ export const getUsers = createAsyncThunk('users/all', async () => {
 	const data = { data: response.data, status: response.status };
 	return data;
 });
-
-export const editProfile = createAsyncThunk(
-	'auth/editProfile',
-	async ({ userData, token }, { rejectWithValue }) => {
-		try {
-			const response = await axios.post(
-				'/api/users/edit',
-				{
-					userData,
-				},
-				{ headers: { authorization: token } }
-			);
-			return { data: response.data, status: response.status };
-		} catch (error) {
-			return rejectWithValue({ status: error.response.status, data: error.response.data });
-		}
-	}
-);
